+++ src/routes/questionRoutes.js
@@ -139,6 +139,12 @@ class QuestionRoutes extends BaseRoute {
       method: "PATCH",
       options: {
         handler: async (request, headers) => {
+          const role = headers.request.auth.artifacts.decoded.role;
+          if (role === "user") {
+            return Boom.forbidden(
+              "Question's update allowed to admin users only"
+            );
+          }
           try {
             const { id } = request.params;
             const { payload } = request;
@@ -193,6 +199,12 @@ class QuestionRoutes extends BaseRoute {
       method: "DELETE",
       options: {
         handler: async (request, headers) => {
+          const role = headers.request.auth.artifacts.decoded.role;
+          if (role === "user") {
+            return Boom.forbidden(
+              "Question's delete allowed to admin users only"
+            );
+          }
           try {
             const { id } = request.params;
             const result = await this.db.delete(id);
