+++ src/routes/userRoutes.js
@@ -25,6 +25,12 @@ class UserRoutes extends BaseRoute {
       method: "POST",
       options: {
         handler: async (request, headers) => {
+          const role = headers.request.auth.artifacts.decoded.role;
+          if (role === "user") {
+            return Boom.forbidden(
+              "User's creation allowed to admin users only"
+            );
+          }
           try {
             let { name, password, role } = request.payload;
             role = role === "admin" ? "admin" : "user";
@@ -68,6 +74,11 @@ class UserRoutes extends BaseRoute {
       method: "GET",
       options: {
         handler: async (request, headers) => {
+          if (role === "user") {
+            return Boom.forbidden(
+              "Retrieve user's information allowed to admin users only"
+            );
+          }
           try {
             const { id } = request.params;
             const result = await this.db.find(id);
@@ -105,6 +116,11 @@ class UserRoutes extends BaseRoute {
       method: "PATCH",
       options: {
         handler: async (request, headers) => {
+          if (role === "user") {
+            return Boom.forbidden(
+              "User's role update allowed to admin users only"
+            );
+          }
           try {
             const { id } = request.params;
             const { payload } = request;
@@ -158,6 +174,9 @@ class UserRoutes extends BaseRoute {
       method: "DELETE",
       options: {
         handler: async (request, headers) => {
+          if (role === "user") {
+            return Boom.forbidden("User's delete allowed to admin users only");
+          }
           try {
             const { id } = request.params;
             const result = await this.db.delete(id);
